requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
Django==3.2.8
djangorestframework==3.12.4
gunicorn==20.1.0
huey==2.4.1
idna==3.3
pycparser==2.20
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
tenacity==8.0.1
urllib3==1.26.7
